+++ local_operator/helpers.py
@@ -15,6 +15,8 @@ import re
 import subprocess
 import sys
 
+from local_operator.types import ResponseJsonSchema
+
 # Configure logging (optional, but helpful for debugging)
 # Use sys.stdout to ensure logs appear if running as a GUI app without a console
 # Note: BasicConfig should ideally be called only once at application entry point.
@@ -290,6 +292,29 @@ def clean_json_response(response_content: str) -> str:
 
     # If we couldn't extract valid JSON, return the original content
     return response_content.strip()
+
+
+def process_json_response(response_str: str) -> ResponseJsonSchema:
+    """Process and validate a JSON response string from the language model.
+
+    Args:
+        response_str (str): Raw response string from the model, which may be wrapped in
+            markdown-style JSON code block delimiters (```json) or provided as a plain JSON object.
+
+    Returns:
+        ResponseJsonSchema: Validated response object containing the model's output.
+            See ResponseJsonSchema class for the expected schema.
+
+    Raises:
+        ValidationError: If the JSON response does not match the expected schema.
+        ValueError: If no valid JSON object can be extracted from the response.
+    """
+    response_content = clean_json_response(response_str)
+
+    # Validate the JSON response
+    response_json = ResponseJsonSchema.model_validate_json(response_content)
+
+    return response_json
 
 
 def is_marker_inside_json(text: str, marker: str) -> bool:
